requirements.txt
earthaccess
folium
geopandas
h5netcdf
rasterio
shapely
xarray
folium
geopandas
h5netcdf
rasterio
shapely
xarray
scikit-learn
pystac-client